Stabilize CLI docs and break import cycle
