+++ .github/workflows/cli-docs.yml
@@ -56,7 +56,7 @@ jobs:
       - name: Build CLI docs
         run: .tox/cli-docs/bin/python scripts/build_cli_docs.py
       - name: Build prompt data
-        run: .tox/cli-docs/bin/python scripts/build_prompt_data.py
+        run: .tox/cli-docs/bin/python scripts/build_prompt_data.py --output fastapi_code_generator/prompt_data.py
       - name: Commit and push if changed
         if: github.event_name == 'push' || github.event.pull_request.head.repo.full_name == github.repository
         run: |
